Snap oracle capture coords to the 1/256 grid the game actually sampled
Factorio's `MapPosition` is fixed point - `int32 / 256` - so a capture
coordinate that is not a multiple of 1/256 is truncated on the way in and the
game evaluates somewhere the fixture never recorded (#186). Seventeen committed
fixtures carry such positions, all of them the far rings that
`test/oracle/capture.ts` builds as `r * Math.cos(a) + 0.5`.

Every one of those rings was being graded as if the game had sampled where the
fixture said. Snapping the coordinate the way the game does, at read time,
improves all 17 and worsens none:

| fixture (worst residual) | as recorded | snapped |
| --- | --- | --- |
| `oracle-temperature` | 17/26 exact, 5.817e-5 | 26/26 exact, 0 |
| `oracle-elevation-lakes` `distance` | 18/26, 4.639e-3 | 26/26, 0 |
| `oracle-vulcanus-resources` sulfuricAcidPatches | 2.942e-3 | 7.153e-8 (41,100x) |
| `oracle-rock-density` | 8/26, 1.570e-3 | 18/26, 8.508e-8 (18,455x) |
| `oracle-elevation-lakes` elevation | 6/17, 7.372e-3 | 13/17, 3.815e-6 (1,933x) |
| `oracle-moisture` | 15/26, 3.070e-5 | 18/26, 5.960e-8 (515x) |
| `oracle-trees` | 83/442, 9.233e-4 | 118/442, 2.593e-6 (356x) |
| `oracle-elevation-island` | 4/17, 6.673e-3 | 10/17, 2.564e-5 (260x) |
| `oracle-aux` | 10/26, 1.262e-5 | 14/26, 5.960e-8 (212x) |
| `oracle-vulcanus-climate` | 60/122, 4.584e-4 | 70/122, 3.359e-6 (136x) |
| `oracle-trees-controls` | 9/51, 8.646e-4 | 10/51, 1.139e-5 (76x) |
| `oracle-vulcanus-temperature` | 185/434, 1.327e-1 | 189/434, 2.639e-3 (50x) |
| `oracle-vulcanus-elevation` | 200/868, 1.332e-1 | 202/868, 5.234e-3 (25x) |
| `oracle-elevation-nauvis` (+ no-cliff) | 3.922e-3 | 3.856e-4 (10x) |
| `oracle-vulcanus-cracks` | 76/305, 1.853e-3 | 79/305, 2.067e-4 (9x) |
| `oracle-resource-starting` | 4.899 abs | 0.641 abs (7.6x) |
| `oracle-vulcanus-biomes` | 1990/3472, 3.092e-4 | 2024/3472, 4.634e-5 (6.7x) |

Every bound above is retightened to its measured post-snap worst. `tree_small_noise`
becomes bit-exact at all 26 positions, worst 0.

## Why this is a measurement, not a fit

Four controls, each of which could have failed and did not:

- **Truncation toward zero, not flooring** - the two are identical on positive
  numbers, so only the negative coordinates discriminate. On `oracle-temperature`'s
  4 all-positive off-grid rows they agree 4/4; on its 10 rows with a negative
  coordinate, truncation is exact where flooring is not 6 times and flooring never
  wins. Across the Vulcanus and tree fixtures, flooring and rounding are each
  WORSE than no snap at all in several arrays.
- **It is the identity on grid-aligned rows** - zero changes across 18,619
  on-grid comparisons.
- **A wrong grid does worse** - a planted 1/128 snap leaves `oracle-resource-starting`
  at 3.012 where 1/256 gives 0.621.
- **The residual lands where a snap predicts** - for 13 of the 14
  `oracle-vulcanus-resources` arrays the post-snap worst now equals the
  on-grid-only worst to every digit, and every measured displacement falls inside
  `[0, 1/256)`.

Each converted spec carries a `countOffGrid(...)` guard so a future re-capture
cannot leave the snap in place as dead code. `test/temperature.spec.ts` and
`test/distanceFromNearestPoint.spec.ts` were watched failing with the snap
removed, reporting exactly the pre-snap counts (17 and 18).

## Mechanisms this refutes

At least eight comments attributed these residuals to "the f32 coordinate floor"
or to "the deep-field point", and both readings were wrong:

- `(12345.75, 6789.125)` - the deep-field point four comments blamed - is ON the
  1/256 grid, so the snap cannot touch it. It was not the worst point in any of
  the seven climate/cracks arrays before the snap; for vulcanus aux it measures
  exactly 0. `test/aux.spec.ts` had even applied that label to an off-grid ring
  point.
- `docs/noise/vulcanus-resources-NOTES.md` blamed "a ring position's exact float
  representation differ[ing] by a ULP or two". At `354.0533905932738` one f32 ulp
  is 3.052e-5 and one f64 ulp is 5.684e-14, against a measured displacement of
  2.609e-3 - 86x and 4.6e10x too large.

## Also in this commit

- **`randomPenalty` now narrows at the store.** `RandomPenalty::run` widens both
  f32 inputs to double, runs the whole chain in f64, and does `fcvt s5, d5; str
  s5` exactly once - so this is the one op where the house f32-everywhere rule
  gives a wrong port, and its single narrowing was missing. Comparing the raw
  return scored 4/40 at worst 1.668e-5, where the spec's own `Math.fround` had
  been recovering it to 40/40. Its only consumer multiplies the returned value;
  narrowing first changes that product in 1240 of 5840 swept cases. Both new
  assertions were watched failing without the fix.
- **The `distance` array is graded for the first time.** 26 game-captured values
  that no spec read, giving `distanceFromNearestPoint` its first test pinned
  against game output rather than hand-written arithmetic. 26/26 exact, worst 0.
- **`test/oracle/capture.ts` snaps its ring positions**, so a future capture
  cannot repeat this. Note the helper there uses `Math.floor` and is right to:
  producing an on-grid coordinate accepts any rounding, while reproducing what
  the game did to an off-grid one is specifically truncation toward zero.

Closes #258. #255 is repurposed to the 1-4 ulp residual that survives the snap
on moisture and aux, which is NOT the snap (3 of each spec's misses are on-grid)
and NOT #191 (both narrowings were applied and moved the counts not at all).

+++ src/noise/randomPenalty.ts
@@ -21,7 +21,47 @@
  * So the value at a given (x, y) depends on the whole batch and its order - which
  * is why a bare `calculate_tile_properties` probe cannot oracle this in isolation.
  * Callers must supply the batch in the same order the game evaluates it.
+ *
+ * ## This op computes in f64 and narrows ONCE - it is the exception to the f32 rule
+ *
+ * Everywhere else in `src/noise/` the rule is f32 after every operation (see
+ * `eval/f32.ts`). Here it would produce a WRONG answer. `RandomPenalty::run`
+ * widens both f32 inputs to double, runs the whole chain in double, and narrows
+ * a single time at the store:
+ *
+ *   +348  ldr   s6, [x11, x8, lsl #2]   // source, f32 in the register buffer
+ *   +352  fcvt  d5, s6                  // widened to DOUBLE
+ *   +416  ucvtf d6, w14                 // the u32 draw -> DOUBLE
+ *   +424  fmul  d6, d6, d7              // * -2^-32 (a DOUBLE constant, 0xBDF0...)
+ *   +432  fcvt  d7, s7                  // amplitude, f32 constant -> DOUBLE
+ *   +436  fmul  d6, d6, d7              // * amplitude, in DOUBLE
+ *   +440  fadd  d5, d5, d6              // + source, in DOUBLE
+ *   +328  fcvt  s5, d5                  // narrowed to f32 exactly once
+ *   +332  str   s5, [x10, x8, lsl #2]   // and stored as f32
+ *
+ * So a Rust port must use `f64` internally and cast to `f32` on the way out.
+ * Writing this one in f32 throughout is the mistake this comment exists to stop.
+ *
+ * The `f32` on the store is load-bearing and was missing until 2026-08-18: 36 of
+ * the fixture's 40 values are not f32 without it, worst gap 1.668e-5, and the
+ * only consumer (`resources/regularPatches.ts`) multiplies the returned value.
+ * Narrowing first changes that product in 1240 of 5840 swept cases, worst 1.19e-7
+ * relative. `test/randomPenalty.spec.ts` asserts the return is f32 directly, so
+ * removing the narrowing goes red rather than being absorbed by a bound.
+ *
+ * Two narrowings the binary also does are deliberately NOT reproduced, because
+ * nothing can currently observe them and an unobservable change is
+ * indistinguishable from a mistake (the rule #191 sets out):
+ *
+ * - `source` is read as f32 at +348. Every shipped source value is f32-exact.
+ * - `amplitude` is read as f32 at +428. `random_penalty_between(min, max, 1)`
+ *   gives 2-0.25, 1-1 and 4-2 across the whole resource catalog, and
+ *   `random_penalty_at(6, 1)` gives 6 - all f32-exact. Only
+ *   `random_penalty_inverse`, whose amplitude is `1/penalty`, could produce a
+ *   non-f32 amplitude, and nothing in base or space-age calls it. Measured: at
+ *   amplitude 1/3 the two readings differ on 1 of 8 outputs by 5.96e-8.
  */
+import { f32 } from "./eval/f32";
 import { seededState, taus88Next } from "./taus88";
 
 /** 2^32, the normalization the binary applies (int32 draw * 2^-32 -> [0,1)). */
@@ -71,8 +111,12 @@ export function randomPenaltyBatch(
     const s = source[i];
     if (s > 0) {
       const u = taus88Next(st) / TWO_POW_32;
-      out[i] = s - amplitude * u;
+      // The chain above is f64 on purpose (see the header). The `f32` here is
+      // the op's single narrowing - `fcvt s5, d5` at +328, then `str s5`.
+      out[i] = f32(s - amplitude * u);
     } else {
+      // The pass-through path stores `source` unchanged, and `source` was read
+      // from an f32 register slot, so this needs no narrowing of its own.
       out[i] = s;
     }
   }
